requirements.txt
Django==4.1.3
tweepy
textblob
pandas
numpy
matplotlib
django-environ
psycopg2
django-crispy-forms
wordcloud
requests
requests-oauthlib
twarc